feat: standalone Data Browser page + clearer Tools nav names
Build on the diagnostics snapshot cache:
- Add a standalone Data Browser page (own route + top-level Tools nav
  item) that reuses the cached snapshot/worker, so it opens instantly.
  Extract the load lifecycle into a shared useDiagnosticsSnapshot hook
  and a shared SnapshotReloadControls component.
- Remove the now-redundant Data Browser tab from Budget File Health.
- Add an Overview card for Data Browser.
- Rename nav items for clarity: 'Budget Diagnostics' -> 'Budget File
  Health', 'ActualQL' -> 'ActualQL Queries' (drop the 'dev' badge).
  Align page titles and README/FEATURES.
