+++ src/app/(app)/budget-diagnostics/page.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from "next";
 import { BudgetDiagnosticsClient } from "./BudgetDiagnosticsClient";
 
 export const metadata: Metadata = {
-  title: "Budget Diagnostics — Actual Bench",
+  title: "Budget File Health — Actual Bench",
 };
 
 export default function BudgetDiagnosticsPage() {
