+++ src/app/(app)/data-browser/DataBrowserClient.tsx
@@ -1,0 +1,25 @@
+"use client";
+
+import dynamic from "next/dynamic";
+
+const DataBrowserView = dynamic(
+  () =>
+    import("@/features/budget-diagnostics/components/DataBrowserView").then(
+      (mod) => mod.DataBrowserView
+    ),
+  {
+    ssr: false,
+    loading: () => (
+      <main className="flex min-h-0 flex-1 flex-col overflow-hidden">
+        <div className="border-b border-border px-6 py-4">
+          <div className="h-7 w-56 animate-pulse rounded-md bg-muted" />
+          <div className="mt-2 h-4 w-96 max-w-full animate-pulse rounded-md bg-muted/70" />
+        </div>
+      </main>
+    ),
+  }
+);
+
+export function DataBrowserClient() {
+  return <DataBrowserView />;
+}
